refactor(mcp): split formatters.ts by domain
Split 1852-line monolithic formatters.ts into domain-specific modules:

- utils.ts - shared grounding formatters
- concept.ts - search/concept/connection formatters
- job.ts - job list/status formatters
- ingest.ts - inspect/ingest file/directory formatters
- system.ts - database/health/allowlist formatters
- epistemic.ts - epistemic status formatters (ADR-065)
- source.ts - source search/polarity axis formatters
- document.ts - document formatters (ADR-084)
- graph.ts - graph CRUD formatters (ADR-089)
- index.ts - barrel re-export for backwards compatibility

No functional changes - pure structural refactoring.
